Add getters for P, I and D terms
By making available the P, I and D terms calculated during the last
active Compute() call, users will be able to log and analyze them to
see how each of the three terms contributed to the PID's output.  This
might be a helpful for tuning the PID parameters.  Well, at least it
was great help to me.